Improve bash logic & add more helpful messages
Changed all if statements to use direct && or || logic instead as I find
this to be more concise and consistent with shell scripting. Also added
success and fail messages many lines of code while hiding the real
stdout and stderr to provide a more user friendly output for easier
debugging. Converted many conditional `[` (AKA `test`) to `[[` to reduce
the potential for future bugs, even if it may reduce compatibility of
running many of these scripts on machines without bash.