Improved the vm script
Added flags for opening with virt-viewer instead of virt-manager and
to optionally open in fullscreen with virt-viewer. Updated the help page
with the new information, and improved the style, along with a new error
codes section. Added a check at the start of the script to ensure required
packages are installed: `virt-manager` and `virsh`. Moved is-running logic
to the `check_is_running` function, and added a check for if the VM exists
before checking if it's running.