vm: manually saving and restoring state files
Added the ability to save state files to a manual location instead of a
managed location like normal. Can then restore that file with the start
command with that file path passed after -f. At the moment no other
VMs or flags can be passed by manually saving or restoring. This will
have to change in the future.