Added fzf, improved variable management, and more
Integrated fzf into all of the commands that take in VM names as
argument except for `is-running` as that command is suppose to be used
by script, not humans. "Improved" variable management by using
`local` instead of `unset` to keep unecessary variables out of global
scope, and by adding `declare -g <var>="" <var>=""` to functions that
export variables into the global scope or change existing variables in
the global scope. Added a --destroy command to shutdown for instantly
killing VMs that aren't responding to the normal shutdown request.