Add the early version of my internet script
This new internet script is a replacement for my existing wi script for
connecting to networks. It's still needs lots of work, but the idea of
this new script is forcing the user to have intention behind there
internet use, needing to specify a reason and a time limit for the
connection. There is also a daily time limit that can be changed in the
config file using 'internet config', to force you to use your time on
the internet sparingly.