Stricter required packages check and more
check_packages now not only checks for nmcli and fzf, but also grep,
awk, systemctl, bash, ps, etc. It also checks that NetworkManager can be
started by systemctl and exits if not.

Changed all references the scripts name: 'internet' to reference the
variable SCRIPT_NAME instead, incase of a change in the future.

Added an 'edit_config' function to check that an EDITOR variable is
stored, and if it isn't then the config file path is printed for the
user to edit manually.